+++ src/test/java/chessgame/domain/piece/KingTest.java
@@ -1,7 +1,5 @@
-package chessgame.domain;
+package chessgame.domain.piece;
 
-import chessgame.domain.piece.Coordinate;
-import chessgame.domain.piece.King;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
