+++ src/test/java/chessgame/domain/piece/PawnTest.java
@@ -1,9 +1,5 @@
-package chessgame.domain;
+package chessgame.domain.piece;
 
-import chessgame.domain.piece.BlackPawn;
-import chessgame.domain.piece.Coordinate;
-import chessgame.domain.piece.Pawn;
-import chessgame.domain.piece.WhitePawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
